+++ app/controllers/api/v1/action_articles_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Api::V1::ActionArticlesController < ApplicationController
+class Api::V1::ActionArticlesController < APIApplicationController
   def index
     render json: { data: Article.where(action_id: params[:id]) }, status: :ok
   end