+++ app/controllers/api/v1/campaign_offers_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Api::V1::CampaignOffersController < ApplicationController
+class Api::V1::CampaignOffersController < APIApplicationController
   def index
     render json: { data: Offer.where(campaign_id: params[:id]) }, status: :ok
   end