+++ config/routes.rb
@@ -17,6 +17,7 @@
 # headers['token'] = '883737733'
 
 Rails.application.routes.draw do
+  root to: 'application#index'
   namespace :api do
     namespace :v1 do
       get '/quizzes', to: 'quiz#index'
@@ -75,5 +76,9 @@ Rails.application.routes.draw do
       post '/app_released', to: 'app_version#check'
     end
   end
+  
+  devise_for :admins
+  mount RailsAdmin::Engine => '/admin', as: 'rails_admin'
+  
   get '/reports', to: 'reports#index'
 end